agregue validacion a rutas y test a las validaciones
