+++ controllers/organization.controller.js
@@ -20,7 +20,7 @@ const findOrganizationById = async function (req, res) {
     const id = req.params.id;
     try {
         const organizationData = await Organization.findByPk(id)
-        if(!organizationData) throw new Error("Organization not found")
+        if(!organizationData) throw new Error("Organization not found", 404)
         const { name, image, phone, address, welcomeText, facebook, linkedin, instagram } = organizationData;
         res.status(200).json({
             name: name,
@@ -41,7 +41,7 @@ const updateOrganization= async (req, res)=>{
     try{
         const organization= await Organization.update(req.body, {where:{id:req.params.id}})                
         if(organization[0])return res.status(200).json(req.body)
-        throw new Error ("Organization not found")
+        throw new Error ("Organization not found", 404)
     }catch(err){
         return res.status(404).json({message:err.message})
     }
